requirements.txt
flask==2.2.5
boto3==1.17.106
google-cloud-storage==1.42.0
python-dotenv==0.15.0
gunicorn==20.0.4
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0

Pillow~=10.0.0
PyQt6~=6.5.1